Use a YAML alias to reduce redundancy.
Instead of repeating the entire "Strawberry" stage multiple times, once fore
each Perl, make a YAML alias to the first instance, and then simply include
the alias and add the PER_VERSION environment variable for subsequent
instances of the stage. A little opaque to those unfamiliar with the YAML
anchor/alias feature (as I was), but definitely cuts down on clutter.

To better demonstrate multiple uses, add a third Strawberry stage that tests a
third version of Perl. Use a variant of 5.26 again, because 5.24 has issues
(see #1 for details).

While at it, switch from URI::db to CPAN::Test::Dummy::Perl5::Make as the
sample dependency, because it is a no-op with no dependencies of its own, so
it keeps things simpler.